+++ src/config/config.ts
@@ -9,14 +9,17 @@ export namespace customConfig {
 		confirmations: number
 	}
 
-	class Crawler {
+	export class CrawlerConfig {
 		enable: boolean
 		txType: string[]
 		callback: string
 		fromBlock: number
+		toBlock: number
 		maxBlockRange?: number
 		pushJobIntervals?: number
 		executeJobConcurrency?: number
+		keepRunning?: boolean
+		forceUpdate?: boolean
 		db: string
 	}
 
@@ -32,7 +35,7 @@ export namespace customConfig {
 		table: string
 	}
 
-	class Database {
+	class DatabaseConfig {
 		postgres: PostgresqlConfig
 		mysql: MysqlConfig
 		sqlite: SqliteConfig
@@ -41,8 +44,8 @@ export namespace customConfig {
 	export class CustomConfig {
 		apiResponseCode: any
 		tx: TxConfig
-		crawler: Crawler
-		database?: Database
+		crawler: CrawlerConfig
+		database?: DatabaseConfig
 	}
 
 	let customConfig: CustomConfig;
@@ -61,7 +64,7 @@ export namespace customConfig {
 	}
 
 	// GetCrawler returns crawler config
-	export function GetCrawler(): Crawler {
+	export function GetCrawler(): CrawlerConfig {
 		if (customConfig.crawler) {
 			return customConfig.crawler;
 		}