+++ src/crawler/db/postgres.ts
@@ -1,8 +1,9 @@
 import {Sequelize} from 'sequelize';
 import {Postgresql} from '@jovijovi/pedrojs-pg';
 import {customConfig} from '../../config';
-import {ICompactTx, ModelAttrs, ModelName, TableName} from './model';
+import {ICompactTx, ModelAttrsMapper, ModelName, TableName} from './model';
 import {Database} from './interface';
+import {DatabaseType} from './constants';
 
 export class PostgresDB extends Database {
 	private _engine: Sequelize;
@@ -28,7 +29,7 @@ export class PostgresDB extends Database {
 		};
 
 		// Define model
-		this.ModelTx = e.define<ICompactTx>(ModelName, ModelAttrs, opts);
+		this.ModelTx = e.define<ICompactTx>(ModelName, ModelAttrsMapper.get(DatabaseType.Postgres), opts);
 
 		// Creates the table if it doesn't exist (and does nothing if it already exists)
 		await this.ModelTx.sync();